+++ crates/editor/src/common/matcher.rs
@@ -134,12 +134,12 @@ impl Matcher {
     }
 }
 
-fn score(choice: &Arc<Choice>, ranges: &[Range<usize>]) -> u32 {
+fn score(choice: &Arc<Choice>, ranges: &[Range<usize>]) -> usize {
     if let Some(n) = choice.number() {
         return n;
     }
     // Closest match first
-   ranges.first().map(|f| f.start as u32).unwrap_or(0)
+   ranges.first().map(|f| f.start).unwrap_or(0)
 }
 
 fn with_case_sensitivity(opt: &str, case_sensitive: bool) -> Cow<str> {
